requirements.txt
flask==3.1.0
numpy==1.26.4
pandas==2.2.3   
pillow==10.4.0
torch==2.5.1
opencv-python==4.10.0